Add files via upload
access to databank, extracts speficit datasets to python. The login and password are required, and of course here replaced with random numbers. The technique of ETL is replicable and consistant.  The wide use of the codes dominantly in Credit Rating Analysis, Credit Risk Management, PD modelling, backtesting, calibrating and other related processes.